third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc-invalid.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that in case of an invalid src, the invalid string is displayed as currentSrc');
</script>
<body>
<img id=invalid src="//]">
<script>
    window.addEventListener("load", t.step_func(function() {
        var img = document.getElementById("invalid");
        assert_equals(img.currentSrc, "//]");
        t.done();
    }));
</script>
</body>
